moves and renames messaging
