+++ gui/src/controls/mod.rs
@@ -2,5 +2,4 @@ mod tweak_control;
 mod next_round_control;
 mod game_status_display;
 mod resource_control;
-mod messageing;
 mod components;
