+++ gui/src/messaging.rs
